QuranAppBackend/requirements.txt
adrf==0.1.2
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-property==0.2.2
asyncio==3.4.3
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
defusedxml==0.7.1
distro==1.8.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.7
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
oauthlib==3.2.2
openai==1.35.14
psycopg[binary]==3.1.14
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0.post1
whitenoise==6.6.0
